refactor: create new ValueReader and ValueWriter
