+++ src/main/java/org/hisp/dhis/smscompression/utils/ValueUtil.java
@@ -1,281 +1,15 @@
 package org.hisp.dhis.smscompression.utils;
 
-import java.text.ParseException;
-
-/*
- * Copyright (c) 2004-2019, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.hisp.dhis.smscompression.SMSCompressionException;
 import org.hisp.dhis.smscompression.SMSConsts;
-import org.hisp.dhis.smscompression.SMSConsts.MetadataType;
 import org.hisp.dhis.smscompression.SMSConsts.ValueType;
-import org.hisp.dhis.smscompression.models.SMSAttributeValue;
-import org.hisp.dhis.smscompression.models.SMSDataValue;
-import org.hisp.dhis.smscompression.models.SMSMetadata;
 import org.hisp.dhis.smscompression.models.SMSValue;
 
 public class ValueUtil
 {
-    // Complex types
-
-    public static void writeAttributeValues( List<SMSAttributeValue> values, boolean useHashing, SMSMetadata meta,
-        BitOutputStream outStream )
-        throws SMSCompressionException
-    {
-        int attributeBitLen = 0;
-
-        writeBool( useHashing, outStream );
-        if ( useHashing )
-        {
-            attributeBitLen = IDUtil.getBitLengthForList( meta.getType( MetadataType.TRACKED_ENTITY_ATTRIBUTE ),
-                MetadataType.TRACKED_ENTITY_ATTRIBUTE );
-            outStream.write( attributeBitLen, SMSConsts.VARLEN_BITLEN );
-        }
-
-        List<SMSValue<?>> smsVals = values.stream().map( v -> v.getSMSValue() ).collect( Collectors.toList() );
-        int fixedIntBitLen = getBitlenLargestInt( smsVals );
-        // We shift the bitlen down one to allow the max
-        outStream.write( fixedIntBitLen - 1, SMSConsts.FIXED_INT_BITLEN );
-
-        for ( Iterator<SMSAttributeValue> valIter = values.iterator(); valIter.hasNext(); )
-        {
-            SMSAttributeValue val = valIter.next();
-            if ( useHashing )
-            {
-                int idHash = BinaryUtils.hash( val.getAttribute(), attributeBitLen );
-                outStream.write( idHash, attributeBitLen );
-            }
-            else
-            {
-                writeString( val.getAttribute(), outStream );
-            }
-            writeSMSValue( val.getSMSValue(), fixedIntBitLen, outStream );
-
-            int separator = valIter.hasNext() ? 1 : 0;
-            outStream.write( separator, 1 );
-        }
-    }
-
-    public static List<SMSAttributeValue> readAttributeValues( SMSMetadata meta, BitInputStream inStream )
-        throws SMSCompressionException
-    {
-        ArrayList<SMSAttributeValue> values = new ArrayList<>();
-        int attributeBitLen = 0;
-        Map<Integer, String> idLookup = null;
-
-        boolean useHashing = readBool( inStream );
-        if ( useHashing )
-        {
-            attributeBitLen = inStream.read( SMSConsts.VARLEN_BITLEN );
-            idLookup = IDUtil.getIDLookup( meta.getType( MetadataType.TRACKED_ENTITY_ATTRIBUTE ), attributeBitLen );
-        }
-        int fixedIntBitLen = inStream.read( SMSConsts.FIXED_INT_BITLEN ) + 1;
-
-        for ( int valSep = 1; valSep == 1; valSep = inStream.read( 1 ) )
-        {
-            String id;
-            if ( useHashing )
-            {
-                int idHash = inStream.read( attributeBitLen );
-                id = idLookup.get( idHash );
-                if ( id == null )
-                    // TODO: Should we be warning and is this the best way to do
-                    // it?
-                    System.out.println(
-                        "WARNING: SMSCompression(readAttributeValues) - Cannot find UID in submission for attribute value" );
-            }
-            else
-            {
-                id = readString( inStream );
-            }
-
-            SMSValue<?> smsValue = readSMSValue( fixedIntBitLen, inStream );
-            values.add( new SMSAttributeValue( id, smsValue ) );
-        }
-
-        return values;
-    }
-
-    public static Map<String, List<SMSDataValue>> groupDataValues( List<SMSDataValue> values )
-    {
-        HashMap<String, List<SMSDataValue>> map = new HashMap<>();
-        for ( SMSDataValue val : values )
-        {
-            String catOptionCombo = val.getCategoryOptionCombo();
-            if ( !map.containsKey( catOptionCombo ) )
-            {
-                ArrayList<SMSDataValue> list = new ArrayList<>();
-                map.put( catOptionCombo, list );
-            }
-            List<SMSDataValue> list = map.get( catOptionCombo );
-            list.add( val );
-        }
-        return map;
-    }
-
-    public static void writeDataValues( List<SMSDataValue> values, boolean useHashing, SMSMetadata meta,
-        BitOutputStream outStream )
-        throws SMSCompressionException
-    {
-        int catOptionComboBitLen = 0;
-        int dataElementBitLen = 0;
-
-        writeBool( useHashing, outStream );
-        if ( useHashing )
-        {
-            catOptionComboBitLen = IDUtil.getBitLengthForList( meta.getType( MetadataType.CATEGORY_OPTION_COMBO ),
-                MetadataType.CATEGORY_OPTION_COMBO );
-            outStream.write( catOptionComboBitLen, SMSConsts.VARLEN_BITLEN );
-
-            dataElementBitLen = IDUtil.getBitLengthForList( meta.getType( MetadataType.DATA_ELEMENT ),
-                MetadataType.DATA_ELEMENT );
-            outStream.write( dataElementBitLen, SMSConsts.VARLEN_BITLEN );
-        }
-
-        List<SMSValue<?>> smsVals = values.stream().map( v -> v.getSMSValue() ).collect( Collectors.toList() );
-        int fixedIntBitLen = getBitlenLargestInt( smsVals );
-        // We shift the bitlen down one to allow the max
-        outStream.write( fixedIntBitLen - 1, SMSConsts.FIXED_INT_BITLEN );
-
-        Map<String, List<SMSDataValue>> valMap = groupDataValues( values );
-
-        for ( Iterator<String> keyIter = valMap.keySet().iterator(); keyIter.hasNext(); )
-        {
-            String catOptionCombo = keyIter.next();
-            if ( useHashing )
-            {
-                int catOptionComboHash = BinaryUtils.hash( catOptionCombo, catOptionComboBitLen );
-                outStream.write( catOptionComboHash, catOptionComboBitLen );
-            }
-            else
-            {
-                writeString( catOptionCombo, outStream );
-            }
-            List<SMSDataValue> vals = valMap.get( catOptionCombo );
-
-            for ( Iterator<SMSDataValue> valIter = vals.iterator(); valIter.hasNext(); )
-            {
-                SMSDataValue val = valIter.next();
-                if ( useHashing )
-                {
-                    int deHash = BinaryUtils.hash( val.getDataElement(), dataElementBitLen );
-                    outStream.write( deHash, dataElementBitLen );
-                }
-                else
-                {
-                    writeString( val.getDataElement(), outStream );
-                }
-                writeSMSValue( val.getSMSValue(), fixedIntBitLen, outStream );
-
-                int separator = valIter.hasNext() ? 1 : 0;
-                outStream.write( separator, 1 );
-            }
-
-            int separator = keyIter.hasNext() ? 1 : 0;
-            outStream.write( separator, 1 );
-        }
-    }
-
-    public static List<SMSDataValue> readDataValues( SMSMetadata meta, BitInputStream inStream )
-        throws SMSCompressionException
-    {
-        int catOptionComboBitLen = 0;
-        int dataElementBitLen = 0;
-        Map<Integer, String> cocIDLookup = null;
-        Map<Integer, String> deIDLookup = null;
-
-        boolean useHashing = readBool( inStream );
-        if ( useHashing )
-        {
-            catOptionComboBitLen = inStream.read( SMSConsts.VARLEN_BITLEN );
-            cocIDLookup = IDUtil.getIDLookup( meta.getType( MetadataType.CATEGORY_OPTION_COMBO ),
-                catOptionComboBitLen );
-
-            dataElementBitLen = inStream.read( SMSConsts.VARLEN_BITLEN );
-            deIDLookup = IDUtil.getIDLookup( meta.getType( MetadataType.DATA_ELEMENT ), dataElementBitLen );
-        }
-        int fixedIntBitLen = inStream.read( SMSConsts.FIXED_INT_BITLEN ) + 1;
-        ArrayList<SMSDataValue> values = new ArrayList<>();
-
-        for ( int cocSep = 1; cocSep == 1; cocSep = inStream.read( 1 ) )
-        {
-            String catOptionCombo;
-            if ( useHashing )
-            {
-                int catOptionComboHash = inStream.read( catOptionComboBitLen );
-                catOptionCombo = cocIDLookup.get( catOptionComboHash );
-                if ( catOptionCombo == null )
-                    // TODO: Should we be warning and is this the best way to do
-                    // it?
-                    System.out.println(
-                        "WARNING: SMSCompression(readDataValues) - Cannot find UID in submission for category option combo" );
-            }
-            else
-            {
-                catOptionCombo = readString( inStream );
-            }
-
-            for ( int valSep = 1; valSep == 1; valSep = inStream.read( 1 ) )
-            {
-                String dataElement;
-                if ( useHashing )
-                {
-                    int dataElementHash = inStream.read( dataElementBitLen );
-                    dataElement = deIDLookup.get( dataElementHash );
-                    if ( dataElement == null )
-                        // TODO: Should we be warning and is this the best way
-                        // to do it?
-                        System.out.println(
-                            "WARNING: SMSCompression(readDataValues) - Cannot find UID in submission for data element" );
-                }
-                else
-                {
-                    dataElement = readString( inStream );
-                }
-
-                SMSValue<?> smsValue = readSMSValue( fixedIntBitLen, inStream );
-                values.add( new SMSDataValue( catOptionCombo, dataElement, smsValue ) );
-            }
-        }
-
-        return values;
-    }
-
-    // Simple types
-
     // TODO: Find a better way to pass down fixedIntBitLen
     public static void writeSMSValue( SMSValue<?> val, int fixedIntBitlen, BitOutputStream outStream )
         throws SMSCompressionException
@@ -431,52 +165,6 @@ public class ValueUtil
         }
         while ( true );
         return s;
-    }
-
-    public static SMSValue<?> asSMSValue( String value )
-    {
-        // BOOL
-        if ( value.equals( "true" ) || value.equals( "false" ) )
-        {
-            Boolean valBool = value.equals( "true" ) ? true : false;
-            return new SMSValue<Boolean>( valBool, ValueType.BOOL );
-        }
-
-        // DATE
-        try
-        {
-            Date valDate = SMSConsts.SIMPLE_DATE_FORMAT.parse( value );
-            return new SMSValue<Date>( valDate, ValueType.DATE );
-        }
-        catch ( ParseException e )
-        {
-            // not a date
-        }
-
-        // INT
-        try
-        {
-            Integer valInt = Integer.parseInt( value );
-            return new SMSValue<Integer>( valInt, ValueType.INT );
-        }
-        catch ( NumberFormatException e )
-        {
-            // not an integer
-        }
-
-        // FLOAT
-        try
-        {
-            Float valFloat = Float.parseFloat( value );
-            return new SMSValue<Float>( valFloat, ValueType.FLOAT );
-        }
-        catch ( NumberFormatException e )
-        {
-            // not a float
-        }
-
-        // If all else fails we can use String
-        return new SMSValue<String>( value, ValueType.STRING );
     }
 
     public static int getBitlenLargestInt( List<SMSValue<?>> values )
